requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
fire==0.5.0
idna==3.4
joblib==1.3.2
lmdb==1.4.1
nltk==3.8.1
numpy==1.25.2
opencv-python==4.6.0.66
pandas==2.0.3
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
regex==2023.8.8
requests==2.31.0
six==1.16.0
termcolor==2.3.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4      